environments/bread2d/toast_quality_optimization_v1.py
"""
RL environment for optimizing toast preparation quality through sequential actions in cutting, toasting, and buttering phases. Agent learns to maximize chef feedback scores by balancing slice quality, toast level, and butter coverage.

Generated from: /Users/nataliechan/Projects/iterate-hack/test_repo_bread2d
Version: 1.0.0
"""

import gymnasium as gym
from gymnasium import spaces
import numpy as np
from typing import Dict, Tuple, Any, Optional


class toast_quality_optimization_v1(gym.Env):
    """RL Environment: RL environment for optimizing toast preparation quality through sequential actions in cutting, toasting, and buttering phases. Agent learns to maximize chef feedback scores by balancing slice quality, toast level, and butter coverage."""
    
    metadata = {
        "render_modes": ["human", "rgb_array"],
        "render_fps": 30,
    }
    
    def __init__(self, render_mode: Optional[str] = None):
        """Initialize the environment.
        
        Args:
            render_mode: Rendering mode ('human', 'rgb_array', or None)
        """
        super().__init__()
        
        self.render_mode = render_mode
        
        # Define action space
        self.action_space = spaces.MultiDiscrete([11, 21, 11])
        
        # Define observation space
        self.observation_space = spaces.Dict({})  # Default empty dict
        
        # Initialize state
        self.state = None
        self.step_count = 0
        self.max_steps = 3
        
    def reset(self, seed: Optional[int] = None, options: Optional[Dict] = None) -> Tuple[Any, Dict]:
        """Reset the environment to initial state.
        
        Args:
            seed: Random seed for reproducibility
            options: Additional options for reset
            
        Returns:
            observation: Initial observation
            info: Additional information
        """
        super().reset(seed=seed)
        
        # Reset state to initial values
        self.state = {}
        
        self.step_count = 0
        
        observation = self._get_observation()
        info = {"step_count": self.step_count}
        
        return observation, info
    
    def step(self, action: Any) -> Tuple[Any, float, bool, bool, Dict]:
        """Run one timestep of the environment.
        
        Args:
            action: Action to take
            
        Returns:
            observation: New observation after action
            reward: Reward for this step
            terminated: Whether episode terminated (success/failure)
            truncated: Whether episode was truncated (timeout)
            info: Additional information
        """
        self.step_count += 1
        
        # Apply action to environment
        # TODO: Apply action to game state
        # self._apply_action_to_game(action)
        
        # Get new observation
        observation = self._get_observation()
        
        # Calculate reward
        reward = 0.0  # TODO: Implement actual reward calculation
        
        # Check termination conditions
        terminated, truncated = (self._check_success() or self._check_failure(), self.step_count >= self.max_steps)
        
        info = {
            "step_count": self.step_count,
            "max_steps": self.max_steps,
        }
        
        return observation, reward, terminated, truncated, info
    
    def render(self):
        """Render the environment."""
        if self.render_mode == "human":
            # Implement human-readable rendering
            print(f"Step: {self.step_count}, State: {self.state}")
        elif self.render_mode == "rgb_array":
            # Return RGB array for visualization
            # TODO: Implement actual rendering based on codebase
            return np.zeros((64, 64, 3), dtype=np.uint8)
    
    def close(self):
        """Clean up resources."""
        pass
    
    def _get_observation(self) -> Any:
        """Get current observation based on state.
        
        Returns:
            observation: Current observation
        """
        return self.state.copy()
    
    # TODO: Implement actual game logic integration
    # These methods need to be implemented based on the codebase:
    # - _apply_action_to_game(action): Apply action to actual game code
    # - _get_game_state(): Extract state from game
    # - _check_game_termination(): Check if game has ended
    
    def _check_success(self) -> bool:
        """Check if episode succeeded.
        
        Returns:
            bool: True if success condition met
        """
        # TODO: Implement success condition based on game
        return False
    
    def _check_failure(self) -> bool:
        """Check if episode failed.
        
        Returns:
            bool: True if failure condition met
        """
        # TODO: Implement failure condition based on game
        return False
