requirements.txt
anthropic>=0.34.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
numpy>=1.20.0
